Added authentication for fileserver

+++ pkg/fileserver/fileserver.go
@@ -6,24 +6,32 @@ import (
 	"log"
 	"os"
 	"path"
+	"strings"
 
 	"github.com/pkg/errors"
 
 	syscall "golang.org/x/sys/unix"
 
 	api "github.com/Mexator/Go-vno/pkg/api/fileserver"
-)
-
-const (
-	serverCatalogPerms os.FileMode = 0777
+	"google.golang.org/grpc/peer"
 )
 
 // FileServer implements FileServerServer
 type FileServer struct {
 	// Absolute path to folder
 	storagePath string
-	storageDir  *os.File
+	// For authentication
+	nsUrl      string
+	storageDir *os.File
 }
+
+const (
+	serverCatalogPerms os.FileMode = 0777
+)
+
+var (
+	ErrNotNS = errors.New("Not a name server")
+)
 
 // Check if server catalog exists and has proper access rights
 func initializeServerCatalog(path string) error {
@@ -60,7 +68,7 @@ func initializeServerCatalog(path string) error {
  * MakeFileServer creates FileServer reading its configuration from
  * `config.json` file.
  */
-func MakeFileServer(storagePath string) (FileServer, error) {
+func MakeFileServer(storagePath, nsUrl string) (FileServer, error) {
 	err := initializeServerCatalog(storagePath)
 	if err != nil {
 		return FileServer{}, errors.Wrap(err,
@@ -68,7 +76,18 @@ func MakeFileServer(storagePath string) (FileServer, error) {
 	}
 
 	file, _ := os.Open(storagePath)
-	return FileServer{storagePath, file}, nil
+	nsUrl = nsUrl[:strings.LastIndex(nsUrl, ":")]
+	return FileServer{storagePath, nsUrl, file}, nil
+}
+
+func (server *FileServer) isNS(ctx context.Context) bool {
+	peer, ok := peer.FromContext(ctx)
+	if !ok {
+		return false
+	}
+	addr := peer.Addr.String()
+	addr = addr[:strings.LastIndex(addr, ":")]
+	return addr == server.nsUrl
 }
 
 // Size returns size of file in request, or error
@@ -130,6 +149,9 @@ func (server *FileServer) Create(
 	ctx context.Context,
 	request *api.CreateRequest,
 ) (*api.CreateResponse, error) {
+	if !server.isNS(ctx) {
+		return nil, errors.Wrap(ErrNotNS, "Failed to create")
+	}
 	filePath := path.Join(server.storagePath, request.Inode)
 
 	_, err := os.Stat(filePath)
@@ -152,6 +174,9 @@ func (server *FileServer) Remove(
 	ctx context.Context,
 	request *api.RemoveRequest,
 ) (*api.RemoveResponse, error) {
+	if !server.isNS(ctx) {
+		return nil, errors.Wrap(ErrNotNS, "Failed to remove")
+	}
 	filePath := path.Join(server.storagePath, request.Inode)
 	err := os.Remove(filePath)
 	if err != nil {
